+++ custom_components/solarcharger/state_machine/solar_charge_state.py
@@ -1,7 +1,10 @@
+# ruff: noqa: TID252
 """State machine interface."""
 
 from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING
+
+from ..const import RunState
 
 if TYPE_CHECKING:
     from .solar_charge import SolarCharge
@@ -24,14 +27,14 @@ class SolarChargeState(ABC):
         self._context = context
 
     @property
-    def state_name(self) -> str:
-        """Get the state name."""
-        return self._state_name
+    def state(self) -> RunState:
+        """Get the run state."""
+        return self._state
 
-    @state_name.setter
-    def state_name(self, name: str) -> None:
-        """Set the state name."""
-        self._state_name = name
+    @state.setter
+    def state(self, state: RunState) -> None:
+        """Set the run state."""
+        self._state = state
 
     # ----------------------------------------------------------------------------
     @abstractmethod