+++ custom_components/solarcharger/state_machine/state_end.py
@@ -20,10 +20,10 @@ class StateEnd(SolarChargeState):
         self,
     ) -> None:
         """Initialise machine state."""
-        self.state_name = RunState.STATE_ENDED.value
+        self.state = RunState.STATE_ENDED
 
     # ----------------------------------------------------------------------------
     async def async_activate_state(self) -> None:
         """Start end state."""
 
-        self.solarcharge.set_run_state(self.state_name)
+        self.solarcharge.set_run_state(self.state)