+++ custom_components/solarcharger/state_machine/state_pause.py
@@ -7,7 +7,7 @@ import logging
 
 from ..chargers.chargeable import Chargeable
 from ..chargers.charger import Charger
-from ..const import SENSOR_SHARE_ALLOCATION, RunState
+from ..const import SENSOR_SHARE_ALLOCATION, ChargeStatus, RunState
 from ..model_charge_stats import ChargeStats
 
 # Import Modules, Not Classes: Instead of from machine import StateA, use
@@ -30,7 +30,7 @@ class StatePause(SolarChargeState):
         self,
     ) -> None:
         """Initialise machine state."""
-        self.state_name = RunState.STATE_PAUSED.value
+        self.state = RunState.STATE_PAUSED
 
     # ----------------------------------------------------------------------------
     def _update_pause_stats(
@@ -38,6 +38,7 @@ class StatePause(SolarChargeState):
     ) -> None:
         """Update pause stats."""
 
+        stats.pause_last_duration = paused_duration
         stats.pause_total_duration += paused_duration
         stats.pause_total_count += 1
         stats.pause_average_duration = timedelta(
@@ -49,6 +50,20 @@ class StatePause(SolarChargeState):
         self.solarcharge.set_pause_stats(stats)
 
     # ----------------------------------------------------------------------------
+    async def _turn_off_charger(self, charger: Charger, chargeable: Chargeable) -> None:
+        """Turn off charger."""
+
+        try:
+            await self.solarcharge.async_set_charge_current(charger, 0)
+            await self.solarcharge.async_turn_charger_switch(charger, turn_on=False)
+            await self.solarcharge.async_update_ha(chargeable)
+
+        except Exception as e:
+            _LOGGER.exception(
+                "%s: Failed to pause charge: %s", self.solarcharge.caller, e
+            )
+
+    # ----------------------------------------------------------------------------
     async def _async_pause_charge(
         self, charger: Charger, chargeable: Chargeable
     ) -> None:
@@ -56,71 +71,53 @@ class StatePause(SolarChargeState):
 
         start_time = self.solarcharge.get_local_datetime()
 
-        is_enough_power = None
-        average_allocated_power = 0
-        data_points = 0
-
         assert self.solarcharge.entities.sensors is not None
         self.solarcharge.entities.sensors[SENSOR_SHARE_ALLOCATION].set_state(0)
 
-        try:
-            await self.solarcharge.async_set_charge_current(charger, 0)
-            await self.solarcharge.async_turn_charger_switch(charger, turn_on=False)
-            await self.solarcharge.async_update_ha(chargeable)
+        await self._turn_off_charger(charger, chargeable)
 
-            self.solarcharge.power_allocations = []
+        self.solarcharge.power_allocations = []
 
-            while True:
-                is_sun_trigger = self.solarcharge.is_sun_trigger()
-                (is_sun_above_start_end_elevations, _) = (
-                    self.solarcharge.is_sun_above_start_end_elevation_triggers()
+        # Initialise counts before starting loop
+        self.solarcharge.stats.loop_success_count = 0
+        self.solarcharge.stats.loop_consecutive_fail_count = 0
+        while True:
+            try:
+                self.solarcharge.abort_if_exceed_max_consecutive_failure()
+
+                charge_status = await self.solarcharge.async_get_loop_status(
+                    charger, chargeable, self.solarcharge.machine_state.state
                 )
-
-                is_enough_power, average_allocated_power, data_points = (
-                    self.solarcharge.is_average_allocated_power_more_than_min_workable_power(
-                        self.solarcharge.max_allocation_count,
-                        self.solarcharge.power_allocations,
-                        raise_the_bar=True,
-                    )
-                )
-
-                if (
-                    (is_sun_trigger and not is_sun_above_start_end_elevations)
-                    or (is_enough_power is not None and is_enough_power)
-                    or not self.solarcharge.is_monitor_available_power()
-                ):
+                if charge_status != ChargeStatus.CHARGE_PAUSE:
                     break
+
+                self.solarcharge.stats.loop_success_count += 1
+                self.solarcharge.stats.loop_consecutive_fail_count = 0
 
                 await asyncio.sleep(self.solarcharge.wait_net_power_update)
 
-        except Exception as e:
-            _LOGGER.exception(
-                "%s: Failed to pause charge: %s", self.solarcharge.caller, e
-            )
+            except Exception as e:
+                self.solarcharge.stats.loop_consecutive_fail_count += 1
+                _LOGGER.exception(
+                    "%s: Failed to pause charge: %s", self.solarcharge.caller, e
+                )
+
+            self.solarcharge.stats.loop_total_count += 1
 
         self.solarcharge.entities.sensors[SENSOR_SHARE_ALLOCATION].set_state(1)
 
         end_time = self.solarcharge.get_local_datetime()
         paused_duration = end_time - start_time
 
-        # Only update stats when exit pause state is due to having enough power.
-        if is_enough_power is not None and is_enough_power:
+        # Only update stats when exiting pause state is due to having enough power.
+        if charge_status == ChargeStatus.CHARGE_CONTINUE:
             self._update_pause_stats(self.solarcharge.stats, paused_duration)
-
-        _LOGGER.warning(
-            "%s: paused_duration=%s (is_enough_power=%s, average_allocated_power=%s, data_points=%s)",
-            self.solarcharge.caller,
-            paused_duration,
-            is_enough_power,
-            average_allocated_power,
-            data_points,
-        )
 
     # ----------------------------------------------------------------------------
     async def async_activate_state(self) -> None:
         """Start pause state."""
 
-        self.solarcharge.set_run_state(self.state_name)
+        self.solarcharge.set_run_state(self.state)
         await self._async_pause_charge(
             self.solarcharge.charger, self.solarcharge.chargeable
         )